1-Turned off tempbuf mode (a minor mode deleting buffers automatically when it detects that the bufer is nto yet used) for normal files.
It has annoying side effects with:

-1 flymake-mode turned on
When tempbuf try to deletw thw buffer, an annoying promp "delete this buffer
even if there is a process attached to it? y/n", appearing at random times
while typing in a different buffer: very annoying!

-2 emacsclient
When the buffer that is deleted was opened with emacs client, the emacs
window is minimized automatically. Even if I am typing something in a
different buffer: very annoying!