1-Added settings to avoid startxwin to start an xterm session each time it is called at bash startup on Cygwin.
When calling up startxwin, a file called .startxwinrc is searched and,
if not found, an xterm prompt is started-up by default.
By adding a void .startxwinrc in the home folder (by the intermediary
of the install script install.sh), the contents of this file is used
for the initialization sequence instead of calling xterm.